requirements.txt
websockets==12.0
python-telegram-bot==20.7
pandas==2.1.4
pandas-ta==0.3.14b
numpy==1.26.4
python-dotenv==1.0.0
APScheduler==3.10.4
aiohttp==3.9.1
